Reorganize global replay code
Theoretically, the scenario and character editors could also have replay systems
