+++ src/tools/prefs.win.cpp
@@ -1,6 +1,7 @@
 
 #include "prefs.hpp"
 #include "game/boe.global.hpp"
+#include "tools/replay.hpp"
 #include "ticpp.h"
 #include <map>
 #include <string>
